+++ test/e2e/specs/site-editor/pages.spec.js
@@ -54,7 +54,7 @@ test.describe( 'Pages', () => {
 		// Insert into Page Content using appender.
 		await page
 			.getByRole( 'region', { name: 'Editor footer' } )
-			.getByRole( 'button', { name: 'Post Content' } )
+			.getByRole( 'button', { name: 'Content' } )
 			.click();
 		await editor.canvas
 			.getByRole( 'button', { name: 'Add block' } )
@@ -79,13 +79,18 @@ test.describe( 'Pages', () => {
 
 		// Switch to template editing focus.
 		await editor.openDocumentSettingsSidebar();
+		await expect(
+			page.locator(
+				'.edit-site-page-panels__edit-template-preview iframe'
+			)
+		).toBeVisible();
 		await page
 			.getByRole( 'region', { name: 'Editor settings' } )
 			.getByRole( 'button', { name: 'Edit template' } )
 			.click();
 		await expect(
 			editor.canvas.getByRole( 'document', {
-				name: 'Block: Post Content',
+				name: 'Block: Content',
 			} )
 		).toContainText(
 			'This is the Post Content block, it will display all the blocks in any single post or page.'